Projects, Playmode, Improvements (#7)
* Added Projects
* Added Playmode
* Improved editor layout with EditorActionBar
* Converted static component manager into per-scene component registry
* Improved file watching system
* Improved event system
* Added support for multiple camera passes
* Improved scripting manager and compiler to use project paths

+++ Native/Engine/Include/Asset/AssetHandle.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <limits>
 
 namespace Odyssey
 {
@@ -15,7 +14,7 @@ namespace Odyssey
 			m_Ptr = nullptr;
 		}
 
-		AssetHandle(uint32_t id, T* ptr)
+		AssetHandle(uint64_t id, T* ptr)
 		{
 			m_ID = id;
 			m_Ptr = ptr;
@@ -23,13 +22,13 @@ namespace Odyssey
 
 	public:
 		T* Get() { return m_Ptr; }
-		int32_t GetID() { return m_ID; }
-		bool IsValid() { return m_ID != -1; }
+		uint64_t GetID() { return m_ID; }
+		bool IsValid() { return m_ID != (uint64_t)(-1); }
 		void Reset() { m_ID = -1; m_Ptr = nullptr; }
 
 	private:
 		friend class AssetManager;
-		int32_t m_ID = -1;
+		uint64_t m_ID = (uint64_t)(-1);
 		T* m_Ptr = nullptr;
 	};
 }
