+++ Native/Engine/Include/Core/Resource.h
@@ -5,10 +5,10 @@ namespace Odyssey
 	class Resource
 	{
 	public:
-		uint32_t GetID() { return m_ID; }
-		void SetID(uint32_t id) { m_ID = id; }
+		uint64_t GetID() { return m_ID; }
+		void SetID(uint64_t id) { m_ID = id; }
 
 	protected:
-		uint32_t m_ID;
+		uint64_t m_ID;
 	};
 }
